delete images when deleting products done

+++ controllers/admin.js
@@ -4,6 +4,8 @@ const Product = require('../models/product');
 const product = require('../models/product');
 const {validationResult} = require('express-validator');
 const { default: mongoose } = require('mongoose');
+const fileHelper = require('../util/file')
+
 exports.getAddProduct = (req, res, next) => {
 
   res.render('admin/edit-product', {
@@ -147,6 +149,7 @@ exports.postEditProduct = (req, res, next) => {
       result.price = updatedPrice
       result.description = updatedDescription
       if(image){
+        fileHelper.deleteFile(product.imageUrl)
         //if new file is uploaded then save it. otherwise dont
         result.imageUrl = image.path
       }
@@ -184,16 +187,24 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId
-  Product.deleteOne({_id : productId, userId : req.user._id})
+Product.findById(productId).then(product => {
+if(!product){
+  return next(new Error('Product not found'))
+}
 
-    .then(result => {
-      console.log('deleted product')
-      res.redirect('/admin/products')
+fileHelper.deleteFile(product.imageUrl)
+return Product.deleteOne({_id : productId, userId : req.user._id})
 
-    })
-    .catch(err=> {
-      const error = new Error(err)
-      error.httpStatusCode = 500
-      return next(error)
-    })
+.then(result => {
+  console.log('deleted product')
+  res.redirect('/admin/products')
+
+})
+.catch(err=> {
+  const error = new Error(err)
+  error.httpStatusCode = 500
+  return next(error)
+})
+})
+  
 }
